platform/requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-Mail==0.10.0
Flask-Caching==2.3.0
SQLAlchemy==2.0.36
psycopg[binary]>=3.2.10
PyJWT==2.10.1
gunicorn==23.0.0
python-dotenv==1.0.1
Werkzeug==3.1.3
